requirements.txt
catboost==1.2
Django==4.1.7
django-crispy-forms==2.0
django-environ==0.10.0
django-heroku==0.3.1
django-rest==0.8.7
djangorestframework==3.14.0
djongo==1.3.6
drf-spectacular==0.26.2
generics==6.0.0
gunicorn==20.1.0
Jinja2==2.11.3
numpy==1.22.4
optuna==3.2.0
pandas==2.0.1
psycopg2-binary==2.9.6
psycopg2==2.9.6
requests==2.25.1
scikit-image==0.19.3
scikit-learn==1.1.2
scipy==1.8.1
Werkzeug==1.0.1
WTForms==2.3.3
